Fix video download bug in syndication
When `next-video-page` was migrated from `n-ui` to `dotcom-page-kit`, it
inadvertently broke video download in syndication.

After migration, `o-viewport` was being imported incorrectly and it
caused the function to error before `USER_DATA` was set.

With `n-ui` all code is transpiled to use the CommonJS module format and
then a plugin is used to check for a default export. If it comes across
one, as it does with `o-viewport`, it injects this:

```
exports.default = {listenTo, ...};

module.exports = exports.default;
```

This allowed `o-viewport` to be imported as it was and not error.

`dotcom-page-kit` doesn't do this which is why `o-viewport` wasn't being
imported correctly. 🐿 v2.12.5